+++ src/BDSComponentFactory.cc
@@ -500,14 +500,19 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateRF(G4double currentArcLength
   // note cavity length is not the same as currentArcLength
   G4double cavityLength = element->l * CLHEP::m;
 
-  G4bool buildIncomingFringe = true;
-  if (prevElement) // could be nullptr
+  // use cavity fringe option, includeFringeFields does not affect cavity fringes
+  G4bool buildCavityFringes = BDSGlobalConstants::Instance()->IncludeFringeFieldsCavities();
+
+  G4bool buildIncomingFringe = buildCavityFringes;
+  // only check if trying to build fringes to begin with as this check should only ever turn off fringe building
+  if (prevElement && buildIncomingFringe) // could be nullptr
 	{// only build fringe if previous element isn't another cavity
 		buildIncomingFringe = prevElement->type != ElementType::_RF;
 	}
 
-  G4bool buildOutgoingFringe = true;
-  if (nextElement) // could be nullptr
+  G4bool buildOutgoingFringe = buildCavityFringes;
+  // only check if trying to build fringes to begin with as this check should only ever turn off fringe building
+	if (nextElement && buildOutgoingFringe) // could be nullptr
 	{// only build fringe if next element isn't another cavity
 		buildOutgoingFringe = nextElement->type != ElementType::_RF;
 	}
@@ -551,7 +556,7 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateRF(G4double currentArcLength
   // aperture radius. Default is beam pipe radius / aper1 if a cavity model isn't specified.
   G4double cavityApertureRadius = cavityInfo->irisRadius;
 
-  if (!BDS::IsFinite((*st)["efield"]) || !includeFringeFields)
+  if (!BDS::IsFinite((*st)["efield"]) || !buildCavityFringes)
 	{// ie no rf field - don't bother with fringe effects
 	  delete stIn;
 	  delete stOut;
@@ -827,6 +832,7 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateKicker(KickerType type)
         {
           // overwrite magnet strength with copy of fringe strength. Should be safe as it has the
           // kicker information from copying previously.
+	  delete st;
           st = new BDSMagnetStrength(*fringeStIn);
           // supply the field for a thin kicker field as it is required to calculate the
           // effective bending radius needed for fringe field and poleface effects
@@ -956,6 +962,8 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateKicker(KickerType type)
 
   if(!HasSufficientMinimumLength(element, false))
     {
+      delete fringeStIn;
+      delete fringeStOut;
       // fringe effect applied in integrator so nothing more to do.
       return new BDSMagnet(t,
 			   baseName,
@@ -1658,15 +1666,24 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateRMatrix()
   GMAD::Element* elementNew = new GMAD::Element(*element);
   elementNew->l = (element->l-thinElementLength)/2.0;
 
-  auto parallelTransport1 = CreateMagnet(elementNew, st, BDSFieldType::paralleltransporter, BDSMagnetType::paralleltransporter);
-  auto rmatrix            = CreateThinRMatrix(0, elementName + "_centre");
-  auto parallelTransport2 = CreateMagnet(elementNew, st, BDSFieldType::paralleltransporter, BDSMagnetType::paralleltransporter);
+  BDSAcceleratorComponent* parallelTransport1 = CreateMagnet(elementNew,
+							     st,
+							     BDSFieldType::paralleltransporter,
+							     BDSMagnetType::paralleltransporter);
+  BDSAcceleratorComponent* rmatrix = CreateThinRMatrix(0,
+						       elementName + "_centre");
+  BDSAcceleratorComponent* parallelTransport2 = CreateMagnet(elementNew,
+							     st,
+							     BDSFieldType::paralleltransporter,
+							     BDSMagnetType::paralleltransporter);
 
   const G4String baseName = elementName;
   BDSLine* bLine = new BDSLine(baseName);
   bLine->AddComponent(parallelTransport1);
   bLine->AddComponent(rmatrix);
   bLine->AddComponent(parallelTransport2);
+
+  delete elementNew;
 
   return bLine;
 }
@@ -2253,18 +2270,6 @@ BDSCavityInfo* BDSComponentFactory::PrepareCavityModelInfoForElement(Element con
   return defaultCI;
 }
 
-BDSMagnetStrength* BDSComponentFactory::PrepareCavityStrength(Element const* el,
-							      G4double cavityLength,
-							      G4double currentArcLength) const
-{
-  BDSMagnetStrength* fringeIn  = nullptr;
-  BDSMagnetStrength* fringeOut = nullptr;
-  BDSMagnetStrength* result = PrepareCavityStrength(el, cavityLength, currentArcLength, fringeIn, fringeOut);
-  delete fringeIn;
-  delete fringeOut;
-  return result;
-}
-
 BDSMagnetStrength* BDSComponentFactory::PrepareCavityStrength(Element const*      el,
 							      G4double            cavityLength,
 							      G4double            currentArcLength,
@@ -2423,10 +2428,10 @@ BDSMagnetStrength* BDSComponentFactory::PrepareMagnetStrengthForMultipoles(Eleme
   BDSMagnetStrength* st = new BDSMagnetStrength();
   SetBeta0(st);
   G4double scaling = el->scaling;
-  G4double length  = el->l;
+  (*st)["length"] = el->l * CLHEP::m; // length needed for thin multipoles
   // component strength is only normalised by length for thick multipoles
   if (el->type == ElementType::_THINMULT)
-    {length = 1;}
+    {(*st)["length"] = 1*CLHEP::m;}
   auto kn = el->knl.begin();
   auto ks = el->ksl.begin();
   std::vector<G4String> normKeys = st->NormalComponentKeys();
@@ -2438,7 +2443,7 @@ BDSMagnetStrength* BDSComponentFactory::PrepareMagnetStrengthForMultipoles(Eleme
     {(*st)[*nkey] = scaling * (*kn);}
   for (; ks != el->ksl.end(); ks++, skey++)
     {(*st)[*skey] = scaling * (*ks);}
-  (*st)["length"] = el->l * CLHEP::m; // length needed for thin multipoles
+
   return st;
 }