+++ src/BDSUtilities.cc
@@ -19,6 +19,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSDebug.hh"
 #include "BDSExtent.hh"
 #include "BDSGlobalConstants.hh"
+#include "BDSPhysicalConstants.hh"
 #include "BDSRunManager.hh"
 #include "BDSUtilities.hh"
 
@@ -28,6 +29,8 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "G4TwoVector.hh"
 #include "G4UserLimits.hh"
 #include "G4Version.hh"
+
+#include "CLHEP/Units/SystemOfUnits.h"
 
 #include <algorithm>
 #include <cmath>
@@ -335,7 +338,7 @@ G4bool BDS::Geant4EnvironmentIsSet()
       if (!env_p)
 	{
 	  result = false;
-	  G4cout << "Vairable: \"" << variable << "\" not found." << G4endl;
+	  G4cerr << "Variable: \"" << variable << "\" not found." << G4endl;
 	}
     }
   return result;
@@ -591,4 +594,10 @@ G4bool BDS::VerboseThisEvent(G4int eventIndex,
 			     G4int eventStop)
 {
   return eventIndex >= eventStart && eventIndex < eventStop;
+}
+
+G4double BDS::Rigidity(G4double momentumMagnitude,
+		       G4double charge)
+{
+  return momentumMagnitude / CLHEP::GeV / BDS::cOverGeV / charge;
 }